requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
alembic==1.15.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pydantic-settings==2.9.1
python-jose[cryptography]
passlib[bcrypt]
pytest==8.0.2
pytest-cov==4.1.0
httpx==0.27.0
pydantic[email]==2.11.3
python-multipart==0.0.20
requests==2.32.3
redis==5.0.1

# Gemini API dependencies
google-genai